+++ app/src/main/java/com/example/seeweather/data/model/HourEntity.kt
@@ -4,6 +4,6 @@ import android.net.Uri
 
 class HourEntity(
 	val date: Long = 0L,
-	val tempC: Float = 0f,
+	val tempC: Int = 0,
 	val icon: Uri? = null,
 )
