+++ app/src/main/java/com/example/seeweather/domain/model/DayWeatherModel.kt
@@ -5,10 +5,10 @@ import android.net.Uri
 class DayWeatherModel(
 	val date: Long = 0L,
 	val icon: Uri? = null,
-	val tempMax: Float = 0f,
-	val tempMin: Float = 0f,
+	val tempMax: Int = 0,
+	val tempMin: Int = 0,
 	val textStatus: String = "",
-	val windPower: Float = 0f,
+	val windPower: Int = 0,
 	val sunrise: Long = 0L,
 	val sunset: Long = 0L,
 )
